refactor(p01-t01): extract archive sync runner into shared module
